Add user reducer
