+++ src/reducer/index.js
@@ -1,14 +1,10 @@
-const initialState = {
-  activeTab: 'User'
-};
+import { combineReducers } from 'redux';
+import userReducer from './user.reducer';
+import tabReducer from './tab.reducer';
 
-const tabReducer = function (state = initialState, action) {
-  switch (action.type) {
-    case "SET_ACTIVE_TAB":
-      return Object.assign({}, state, {activeTab: action.data});
-    default:
-      return state;
-  }
-};
+const reducer = combineReducers({
+  tab: tabReducer,
+  user: userReducer
+});
 
-export default tabReducer;
+export default reducer;
