+++ src/reducer/initial-state.js
@@ -1,0 +1,13 @@
+const initialState = {
+  tab: {
+    activeTab: 'User'
+  },
+  user: {
+    email: '',
+    name: '',
+    role: '',
+    password: ''
+  }
+};
+
+export default initialState;
