+++ src/reducer/user.reducer.js
@@ -1,0 +1,10 @@
+import initialState from './initial-state';
+
+const userReducer = function (state = initialState.user, action) {
+  switch (action.type) {
+    default:
+      return state;
+  }
+};
+
+export default userReducer;
